Don't stop music playback for alarms
Alarms are now played over any active music playback.

See also:
- https://github.com/rinigus/pure-maps/issues/120
- https://github.com/rinigus/pure-maps/commit/53707627c4a6f858bf53507c5284424bc1e992da
- https://forum.sailfishos.org/t/playing-audio-file-with-qmediaplayer-pause-media-player/2732/5
- https://github.com/Karry/osmscout-sailfish/commit/84dd52280de9304660d0042d06dde8aa1ea98bf4

I decided to replace Audio with SoundEffect (in QML) because it worked and
didn't require further testing.

Maybe in the future it may be preferable to use Audio with
e.g. NEMO_RESOURCE_CLASS_OVERRIDE=alarm instead because SoundEffect
trades lower latency for worse performance but latency is irrelevant
for this app.